requirements.txt
tldextract
pyenchant
compound-word-splitter
numpy
pandas
dnspython
pyasn
netaddr
flask
networkx